Enable overriding .shebang line with im_prg
Here is a motivational example:

On machines with coreutils older than v8.30 (i.e. Ubuntu 18.04 LTS),
it is impossible to  use the `-S` option on `/usr/bin/env` to split
the shebang line into a command with arguments.
Tricks like

    #/usr/bin/env -S nice -n5 python
    ...

would on these systems not work.

This patch enables the user to instead set use
{.shebang im_prg="nice -n5 python"}